backend/requirements.txt
# Secure Notes API - Backend Dependencies

# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==42.0.5

# Validation
pydantic==2.5.2
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client (for testing)
requests==2.31.0

# WSGI Server
gunicorn==21.2.0

# Rate Limiting
Flask-Limiter==3.5.0
redis==5.0.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==20.1.0

# Code Quality
flake8==6.1.0
black==23.12.1
mypy==1.7.1

# Database Migrations
Flask-Migrate==4.0.5
alembic==1.13.0
